refactor: moving header to subpackage
