+++ node/header/config.go
@@ -1,0 +1,19 @@
+package header
+
+// Config contains configuration parameters for header retrieval
+type Config struct {
+	// TrustedHash is the Block/Header hash that Nodes use as starting point for header synchronization.
+	// Only affects the node once on initial sync.
+	TrustedHash string
+	// TrustedPeers are the peers we trust to fetch headers from.
+	// Note: The trusted does *not* imply Headers are not verified, but trusted as reliable to fetch headers
+	// at any moment.
+	TrustedPeers []string
+}
+
+func DefaultConfig() Config {
+	return Config{
+		TrustedHash:  "",
+		TrustedPeers: make([]string, 0),
+	}
+}
