+++ node/header/opts.go
@@ -1,0 +1,27 @@
+package header
+
+import (
+	"go.uber.org/fx"
+)
+
+type Option func(*settings)
+
+// settings store values that can be augmented or changed for Node with Options.
+type settings struct {
+	cfg  *Config
+	opts []fx.Option
+}
+
+// WithTrustedHash sets TrustedHash to the Config.
+func WithTrustedHash(hash string) Option {
+	return func(sets *settings) {
+		sets.cfg.TrustedHash = hash
+	}
+}
+
+// WithTrustedPeers appends new "trusted peers" to the Config.
+func WithTrustedPeers(addr ...string) Option {
+	return func(sets *settings) {
+		sets.cfg.TrustedPeers = append(sets.cfg.TrustedPeers, addr...)
+	}
+}
