+++ node/services/config.go
@@ -1,27 +1,14 @@
 package services
 
 import (
-	"encoding/hex"
 	"time"
 
 	logging "github.com/ipfs/go-log/v2"
-	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/multiformats/go-multiaddr"
-	tmbytes "github.com/tendermint/tendermint/libs/bytes"
-
-	"github.com/celestiaorg/celestia-node/params"
 )
 
 var log = logging.Logger("node/services")
 
 type Config struct {
-	// TrustedHash is the Block/Header hash that Nodes use as starting point for header synchronization.
-	// Only affects the node once on initial sync.
-	TrustedHash string
-	// TrustedPeers are the peers we trust to fetch headers from.
-	// Note: The trusted does *not* imply Headers are not verified, but trusted as reliable to fetch headers
-	// at any moment.
-	TrustedPeers []string
 	// NOTE: All further fields related to share/discovery.
 	// PeersLimit defines how many peers will be added during discovery.
 	PeersLimit uint
@@ -34,43 +21,8 @@ type Config struct {
 
 func DefaultConfig() Config {
 	return Config{
-		TrustedHash:       "",
-		TrustedPeers:      make([]string, 0),
 		PeersLimit:        3,
 		DiscoveryInterval: time.Second * 30,
 		AdvertiseInterval: time.Second * 30,
 	}
-}
-
-func (cfg *Config) trustedPeers(bpeers params.Bootstrappers) (infos []peer.AddrInfo, err error) {
-	if len(cfg.TrustedPeers) == 0 {
-		log.Infof("No trusted peers in config, initializing with default bootstrappers as trusted peers")
-		return bpeers, nil
-	}
-
-	infos = make([]peer.AddrInfo, len(cfg.TrustedPeers))
-	for i, tpeer := range cfg.TrustedPeers {
-		ma, err := multiaddr.NewMultiaddr(tpeer)
-		if err != nil {
-			return nil, err
-		}
-		p, err := peer.AddrInfoFromP2pAddr(ma)
-		if err != nil {
-			return nil, err
-		}
-		infos[i] = *p
-	}
-
-	return
-}
-
-func (cfg *Config) trustedHash(net params.Network) (tmbytes.HexBytes, error) {
-	if cfg.TrustedHash == "" {
-		gen, err := params.GenesisFor(net)
-		if err != nil {
-			return nil, err
-		}
-		return hex.DecodeString(gen)
-	}
-	return hex.DecodeString(cfg.TrustedHash)
 }
