+++ node/state/module.go
@@ -2,6 +2,7 @@ package state
 
 import (
 	"context"
+	"github.com/celestiaorg/celestia-node/node/header"
 
 	logging "github.com/ipfs/go-log/v2"
 	"go.uber.org/fx"
@@ -10,7 +11,6 @@ import (
 
 	"github.com/celestiaorg/celestia-node/fraud"
 	"github.com/celestiaorg/celestia-node/libs/fxutil"
-	"github.com/celestiaorg/celestia-node/node/services"
 	"github.com/celestiaorg/celestia-node/service/state"
 )
 
@@ -41,7 +41,7 @@ func Module(tp node.Type, cfg Config, options ...Option) fx.Option {
 			fx.Provide(fx.Annotate(state.NewService,
 				fx.OnStart(func(ctx context.Context, lc fx.Lifecycle, fservice fraud.Service, serv *state.Service) error {
 					lifecycleCtx := fxutil.WithLifecycle(ctx, lc)
-					return services.FraudLifecycle(ctx, lifecycleCtx, fraud.BadEncoding, fservice, serv.Start, serv.Stop)
+					return header.FraudLifecycle(ctx, lifecycleCtx, fraud.BadEncoding, fservice, serv.Start, serv.Stop)
 				}),
 				fx.OnStop(func(ctx context.Context, serv *state.Service) error {
 					return serv.Stop(ctx)
