+++ node/tests/sync_test.go
@@ -2,6 +2,7 @@ package tests
 
 import (
 	"context"
+	"github.com/celestiaorg/celestia-node/node/header"
 	"testing"
 	"time"
 
@@ -49,7 +50,7 @@ func TestSyncLightWithBridge(t *testing.T) {
 	addrs, err := peer.AddrInfoToP2pAddrs(host.InfoFromHost(bridge.Host))
 	require.NoError(t, err)
 
-	light := sw.NewLightNode(nodebuilder.WithTrustedPeers(addrs[0].String()))
+	light := sw.NewLightNode(nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 
 	err = light.Start(ctx)
 	require.NoError(t, err)
@@ -98,7 +99,7 @@ func TestSyncStartStopLightWithBridge(t *testing.T) {
 	require.NoError(t, err)
 
 	store := nodebuilder.MockStore(t, nodebuilder.DefaultConfig(node.Light))
-	light := sw.NewNodeWithStore(node.Light, store, nodebuilder.WithTrustedPeers(addrs[0].String()))
+	light := sw.NewNodeWithStore(node.Light, store, nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 	require.NoError(t, light.Start(ctx))
 
 	h, err = light.HeaderServ.GetByHeight(ctx, 30)
@@ -109,7 +110,7 @@ func TestSyncStartStopLightWithBridge(t *testing.T) {
 	require.NoError(t, light.Stop(ctx))
 	require.NoError(t, sw.RemoveNode(light, node.Light))
 
-	light = sw.NewNodeWithStore(node.Light, store, nodebuilder.WithTrustedPeers(addrs[0].String()))
+	light = sw.NewNodeWithStore(node.Light, store, nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 	require.NoError(t, light.Start(ctx))
 
 	h, err = light.HeaderServ.GetByHeight(ctx, 40)
@@ -149,7 +150,7 @@ func TestSyncFullWithBridge(t *testing.T) {
 	addrs, err := peer.AddrInfoToP2pAddrs(host.InfoFromHost(bridge.Host))
 	require.NoError(t, err)
 
-	full := sw.NewFullNode(nodebuilder.WithTrustedPeers(addrs[0].String()))
+	full := sw.NewFullNode(nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 	require.NoError(t, full.Start(ctx))
 
 	h, err = full.HeaderServ.GetByHeight(ctx, 30)
@@ -195,7 +196,7 @@ func TestSyncLightWithFull(t *testing.T) {
 	addrs, err := peer.AddrInfoToP2pAddrs(host.InfoFromHost(bridge.Host))
 	require.NoError(t, err)
 
-	full := sw.NewFullNode(nodebuilder.WithTrustedPeers(addrs[0].String()))
+	full := sw.NewFullNode(nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 	require.NoError(t, full.Start(ctx))
 
 	h, err = full.HeaderServ.GetByHeight(ctx, 30)
@@ -206,7 +207,7 @@ func TestSyncLightWithFull(t *testing.T) {
 	addrs, err = peer.AddrInfoToP2pAddrs(host.InfoFromHost(full.Host))
 	require.NoError(t, err)
 
-	light := sw.NewLightNode(nodebuilder.WithTrustedPeers(addrs[0].String()))
+	light := sw.NewLightNode(nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 
 	err = sw.Network.UnlinkPeers(bridge.Host.ID(), light.Host.ID())
 	require.NoError(t, err)
@@ -259,7 +260,7 @@ func TestSyncLightWithTrustedPeers(t *testing.T) {
 
 	trustedPeers := []string{addrs[0].String()}
 
-	full := sw.NewFullNode(nodebuilder.WithTrustedPeers(addrs[0].String()))
+	full := sw.NewFullNode(nodebuilder.WithHeaderOption(header.WithTrustedPeers(addrs[0].String())))
 	require.NoError(t, full.Start(ctx))
 
 	h, err = full.HeaderServ.GetByHeight(ctx, 30)
@@ -272,7 +273,7 @@ func TestSyncLightWithTrustedPeers(t *testing.T) {
 
 	trustedPeers = append(trustedPeers, addrs[0].String())
 
-	light := sw.NewLightNode(nodebuilder.WithTrustedPeers(trustedPeers...))
+	light := sw.NewLightNode(nodebuilder.WithHeaderOption(header.WithTrustedPeers(trustedPeers...)))
 
 	err = light.Start(ctx)
 	require.NoError(t, err)
